Rename unit-conversion module and add more tests
